Extracted components from AttachableInfo
